blockmodel_to_summary 기능 작성 완료

+++ src/main/java/com/metalbird/beautier/controller/model/BlockSummaryResult.java
@@ -3,7 +3,6 @@ package com.metalbird.beautier.controller.model;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.Map;
 
 @Data
@@ -12,7 +11,7 @@ public class BlockSummaryResult {
     
     private long newestBlockNumber;
     private long txBlockCount;
-    private Map<BigInteger, Integer> orderedPriceTxCntMap;
+    private Map<Double, Integer> orderedPriceTxCntMap;
     private double avgPrice;
     private double maxPrice;
     private double minPrice;
